+++ R/dic_attr.R
@@ -7,18 +7,26 @@
 #' @return Attribute value
 #' @export
 dic_attr <- function(x, var) {
-  out <- attr(x, .opt$dic)
-  out[[var]]
+  out <- attr(x, opt("dic"))
+  if (var %in% names(.opt)) {
+    out[[opt(var)]]
+  } else {
+    out[[var]]
+  }
 }
 
 #' @rdname dic_attr
 #' @param value set value
 #' @export
 "dic_attr<-" <- function(x, var, value) {
-  dic_attr <- attr(x, .opt$dic)
+  dic_attr <- attr(x, opt("dic"))
   if (is.null(dic_attr)) dic_attr <- list()
 
-  dic_attr[[var]] <- value
-  attr(x, .opt$dic) <- dic_attr
+  if (var %in% names(.opt)) {
+    dic_attr[[opt(var)]] <- value
+  } else {
+    dic_attr[[var]] <- value
+  }
+  attr(x, opt("dic")) <- dic_attr
   x
 }