+++ R/rename_item.R
@@ -4,9 +4,9 @@
 #'
 #' @param data A data frame
 #' @param pattern A character string with the syntax of the glue function (see
-#'   example).
-#' @param max_chars,chars If not NULL, only the first n chars og the long label
-#'   will be applied.
+#'   [glue()]).
+#' @param max_chars,chars If not NULL, only the first n chars of the resulting
+#'   label will be applied.
 #' @return A renamed data frame
 #' @examples
 #' ex_itrf  |>